+++ 3DSkit.py
@@ -13,7 +13,7 @@ from io import BytesIO, StringIO
 from util import error
 import util
 
-__version__ = '1.33.73'
+__version__ = '1.33.74'
 
 
 def parse_opts(s):